setup.py: Add LIBGIT2_LIB to override the library search path
I build libgit2 in a `build/` subdirectory of a Git checkout, and
never bother installing it.  I'd like to link pygit2 against it there,
but LIBGIT2 assumes libgit2.so will be in the default library search
path or $LIBGIT/lib.  This patch adds LIBGIT2_LIB, a new environment
variable for overriding the additional library search paths.

An alternative to this patch is to set LDFLAGS="-L..." explicitly
before building pygit2.  I've avoided that approach here minimize the
difference between configuring this library path explicitly, and
configuring implicitly via LIBGIT2.