Improve diff & diff_to_buffer doc formatting
We switch to a parameter list for both functions and add a return type.
We also remove the indentation from the second line of the explanation
which was causing Sphinx to return it as two lines instead of one
continuous one.

We also added return types.

I am not sure of the type of the GIT_DIFF* flags so I have not included
that.