branch: correct notion of remote-tracking branch and upstream
The current documentation text seems to be very confused about what a
remote-tracking branch is and what the relationship to an upstream is.

Correct the text to talk about the upstream branch when dealing with the
upstream configuration, instead of implying that it's related to the
setup of remote-tracking branches.