setup: download and build libgit2
This is a new fallback for when LIBGIT2 is not defined. If libgit2 is not
found, then setup.py will download and build libgit2 so that the correct
version matches pygit2.

This is helpful for users on systems that don't have access to install packages
or systems with a globally installed libgit that is the incorrect version.