Patch: Add __str__ and __bytes__ for undecoded content.
Patch.patch assumes all content to be encoded in UTF-8 and forcefully
replaces any non-decodable sequences. This can lead to corruption for
content that either does not conform to any specific encoding altogether, or
uses an encoding that is incompatible with, or ambinuous to UTF-8.

This change adds __str__(), __bytes__() methods to Patch that return
the unmodified, raw bytes. It also adds a new .decode() method that
gives greater control over how text decoding happens, deprecating
Patch.patch.