Update to the libgit2 master branch to use mailmap
This patch shouldn't be merged until libgit2 0.28 is released, as the
mailmap support is not present outside of the master branch currently.
This patch sets us up to test against the master branch for now.

NOTE: A test fails until https://github.com/libgit2/libgit2/pull/4688 is
landed on libgit2 master.