Deprecate Patch.patch and add Patch.text and Patch.data
This is an attempt to finish up the work in PR #790 originally done by
@erikvanzijst. Thanks to him for the initial work.

Patch.patch assumes all content to be encoded in UTF-8 and forcefully
replaces any non-decodeable sequences. This can lead to corruption for
content that either does not conform to any specific encoding altogether,
or uses an encoding that is incompatible with, or ambiguous to UTF-8.

As discussed in #790, this change deprecates Patch.patch in favor of
Patch.text and adds Patch.data, which returns the unmodified, raw bytes.