ConfigEntry: keep a reference to ConfigIterator
So git_config_iterator_free is only called when all the config entries
are deleted.

Because as stated in https://libgit2.org/libgit2/#HEAD/group/config/git_config_next
"The pointers returned by this function are valid until the iterator is
freed."

Unfortunately this is not enough, through iteration the pointers in the
git_config_entry struct become invalid. So the statement above in
libgit2's documentation looks wrong, the pointers are invalid earlier.