Introduce Refdb type
Currently this allows you to manipulate refdbs in a vacuum. A later
commit will introduce RefdbBackend, then allow you to create a
Repository given a refdb backend and an odb backend.

The refdb is more or less a write-only interface from libgit2, so it's
not really possible to verify that the functions added here have
affected libgit2's state in the desired way. For this reason, tests are
not included for this type.