Add repr, str, == and != support to IndexEntry.
IndexEntry is a simple datatype which can be usefully displayed or
compated to any other IndexEntry.
Not done for larger types Index or ConflictCollection since
its not clear if, for instance, two Indexes can be == if they have
different paths.

Added ConflictIterator.__iter__ - this is required by the iterator
protocol. See
https://docs.python.org/3/library/stdtypes.html#iterator-types