bump cibuildwheel to 2.9.0 to build Python 3.11 wheels, also test on CI
Python 3.11.0rc1 is now released, which is guaranteed to be ABI compatible with the final release.
And the cibuildwheel v2.9.0 release includes Python 3.11 and builds wheels by default.

I also added a CI job to test with Python 3.11 in Linux so that any issues are caught. After final
release, we can remove the Python 3.10 version with 3.11 and reduce matrix that we are testing now.