Fix crash in Signature repr/str with encoding mismatch (#1205)
`to_unicode(..., ..., NULL)` is equivalent to `to_unicode(..., ..., "strict")`. It returns NULL if the string cannot be decoded to the desired encoding.

PyUnicode_FromFormat will crash if we pass NULL pointers to be formatted as %R or %U.

Instead, let's use `to_unicode(..., ..., "replace")` so that we always get a non-NULL object to pass to PyUnicode_FromFormat.

This ties in with #1155.