Improve typing: add References to Repository
pygit2 was removed from the python typeshed, but the built in types of
pygit2 are not complete.

This change tries to improve this situation, by fixing type errors that
occur when activating typing for the tests.
So first the test functions (for the public API) in test/test_refs.py
were adapted, in order to activate mypy type checking.
Then _pygit2.pyi was adapted to fix those mypy type checking errors.

The skeleton was provided by:
    stubgen -m pygit2.references -o /tmp/pygit2

The following command
    mypy test
returned the following summary before this change:
    Found 289 errors in 10 files (checked 50 source files)
and after this change:
    Found 289 errors in 10 files (checked 50 source files)